fiqo-cli/pyproject.toml
[project]
name = "fiqo-cli"
version = "1.0.0"
description = "Fiqo CLI"
authors = [{ name = "Behlul Mansur Cildir", email = "mansurcildir@gmail.com" }]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "typer",
    "questionary",
    "requests",
    "pyjwt",
    "python-dotenv"
]

[project.scripts]
fiqo = "fiqo.main:app"
